--- src/main/java/org/example/email/SendEmail.java
@@ -1,44 +1,0 @@
-package org.example.email;
-
-import javax.mail.*;
-import javax.mail.internet.InternetAddress;
-import javax.mail.internet.MimeBodyPart;
-import javax.mail.internet.MimeMessage;
-import javax.mail.internet.MimeMultipart;
-import java.util.Properties;
-
-public class SendEmail {
-
-    public static void sendEmail(String toEmail, Integer msg) throws Exception {
-        Properties properties = new Properties();
-        properties.put("mail.smtp.auth", "true");
-        properties.put("mail.smtp.port", "587"); // Use the appropriate port for TLS
-        properties.put("mail.smtp.starttls.enable", "true");
-        properties.put("mail.smtp.host", "smtp.gmail.com");
-
-        String username = "kamoliddinabduxalikov02@gmail.com";
-        String password = "rhgp oxrq wmtp owwn";
-        Session session = Session.getInstance(properties, new Authenticator() {
-            @Override
-            protected PasswordAuthentication getPasswordAuthentication() {
-                return new PasswordAuthentication(username, password);
-            }
-        });
-
-        Message message = new MimeMessage(session);
-        message.setFrom(new InternetAddress("kamoliddinabduxalikov02@gmail.com"));
-        message.setRecipients(
-                Message.RecipientType.TO, InternetAddress.parse(toEmail));
-        message.setSubject("Mail Subject");
-
-        MimeBodyPart mimeBodyPart = new MimeBodyPart();
-        mimeBodyPart.setContent(msg.toString(), "text/html; charset=utf-8");
-
-        Multipart multipart = new MimeMultipart();
-        multipart.addBodyPart(mimeBodyPart);
-
-        message.setContent(multipart);
-
-        Transport.send(message);
-    }
-}
