+++ src/main/java/org/example/service/EmailService.java
@@ -1,5 +1,6 @@
 package org.example.service;
 
+import javax.mail.MessagingException;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
@@ -9,4 +10,6 @@ public interface EmailService {
     Map<String,Integer> EMAIL_HISTORY_MAP = Collections.synchronizedMap(new HashMap<>());
 
     void addEmailHistory(String email, Integer msg);
+
+    void sendEmail(String toEmail,Integer msg) throws MessagingException;
 }
